Consolidate Iterator structs as generic
