+++ libnss/src/host.rs
@@ -1,5 +1,4 @@
 use crate::interop::CBuffer;
-use std::collections::VecDeque;
 use std::mem;
 use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
 
@@ -105,31 +104,6 @@ pub struct CHost {
     pub h_addr_list: *mut *mut libc::c_char,
 }
 
-pub struct HostIterator {
-    items: Option<VecDeque<Host>>,
-}
-
-impl HostIterator {
-    pub fn new() -> Self {
-        HostIterator { items: None }
-    }
-
-    pub fn open(&mut self, items: Vec<Host>) {
-        self.items = Some(VecDeque::from(items));
-    }
-
-    pub fn next(&mut self) -> Option<Host> {
-        match self.items {
-            Some(ref mut val) => val.pop_front(),
-            None => panic!("Iterator not currently open"),
-        }
-    }
-
-    pub fn close(&mut self) {
-        self.items = None;
-    }
-}
-
 #[macro_export]
 macro_rules! libnss_host_hooks {
 ($mod_ident:ident, $hooks_ident:ident) => (
@@ -142,23 +116,23 @@ macro_rules! libnss_host_hooks {
             use std::str;
             use std::sync::{Mutex, MutexGuard};
             use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
-            use $crate::interop::{CBuffer, NssStatus};
-            use $crate::host::{CHost, HostHooks, HostIterator, AddressFamily};
+            use $crate::host::{CHost, HostHooks, Host, AddressFamily};
+            use $crate::interop::{CBuffer, NssStatus, Iterator};
 
             lazy_static! {
-            static ref [<HOST_ $mod_ident _ITERATOR>]: Mutex<HostIterator> = Mutex::new(HostIterator::new());
+            static ref [<HOST_ $mod_ident _ITERATOR>]: Mutex<Iterator<Host>> = Mutex::new(Iterator::<Host>::new());
             }
 
             #[no_mangle]
             extern "C" fn [<_nss_ $mod_ident _sethostent>]() -> libc::c_int {
-                let mut iter: MutexGuard<HostIterator> = [<HOST_ $mod_ident _ITERATOR>].lock().unwrap();
+                let mut iter: MutexGuard<Iterator<Host>> = [<HOST_ $mod_ident _ITERATOR>].lock().unwrap();
                 iter.open(super::$hooks_ident::get_all_entries());
                 NssStatus::Success.to_c()
             }
 
             #[no_mangle]
             extern "C" fn [<_nss_ $mod_ident _endhostent>]() -> libc::c_int {
-                let mut iter: MutexGuard<HostIterator> = [<HOST_ $mod_ident _ITERATOR>].lock().unwrap();
+                let mut iter: MutexGuard<Iterator<Host>> = [<HOST_ $mod_ident _ITERATOR>].lock().unwrap();
                 iter.close();
                 NssStatus::Success.to_c()
             }
@@ -166,7 +140,7 @@ macro_rules! libnss_host_hooks {
             #[no_mangle]
             unsafe extern "C" fn [<_nss_ $mod_ident _gethostent_r>](result: *mut CHost, buf: *mut libc::c_char, buflen: libc::size_t,
                                                                   _errnop: *mut libc::c_int) -> libc::c_int {
-                let mut iter: MutexGuard<HostIterator> = [<HOST_ $mod_ident _ITERATOR>].lock().unwrap();
+                let mut iter: MutexGuard<Iterator<Host>> = [<HOST_ $mod_ident _ITERATOR>].lock().unwrap();
                 match iter.next() {
                     None => $crate::interop::NssStatus::NotFound.to_c(),
                     Some(entry) => {
